fix overflow in conversion from long int to int

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -9,7 +9,7 @@
 
 int main(void)
 {
-	int i, j, test, n = 612852475143;
+	long int i, j, test, n = 612852475143;
 
 	for (i = 2; i <= n; i++)
 	{
